Greatly reorganized command line configuration variables.
Now there's a PixiConfig class that holds everything that's passed from the command line, instead of various variables scattered around. 
The various defines that hold the constants to index the paths and extensions arrays are now enum classes with a convenience function to convert to int. 
Due to the changes to how command line params are passed, most of the things are now std::string instead of const char* and most the function parameters went from const char* to std::string_view.