[NoTest] Improvements to API surface. Enable optional static linking.
There are now APIs to query most information off of sprites, as well as parse them from CFG and JSON files.
The API is not the best but it's a WIP. All bindings (C#, Python and C) are updated.
Added a CMake variable that can be specified at CMake generation time (PIXI_LIBRARY_TYPE) which can be either STATIC or SHARED (SHARED by default) to be able to link pixi either as a dinamic or static library.