Add GPLv3 license.
This commit adds a license to the repository.
This wasn't done only by myself, I asked permission from all the major contributors of this repository (me, lx5, etc), including the repository owner, JackTheSpades, and the author of the tool from which the idea for this tool was taken, p4plus2. 
Having a license improves legality around this piece of OSS and makes us able to ship this tool with a statically linked asar library instead of having to ship a .dll.