Remove arguments to replace stdin, stdout and make argv const
The replacing of stdin and stdout using freopen was a bad design decision, mainly because it's not possible to restore the old stdin and stdout after freopen is called. This meant that the application calling pixi_run would get its own stdin and stdout closed after the function finished.
A better design would be requiring two FILE* arguments but that would also require rewriting all of pixi's print statements to use a proper file handler instead of directly going to stdout, which requires a hefty bit of work.
Argv const is just a "why not".