Handle high values of Y in sprite data loader (fixes #70).
If a value of Y of $00FE or higher is present (which can happen when a lot of sprites are in a level)
then it would overflow and roll over, making the loader set a bogus value in !extra_byte_1/2/3
for the sprite extra bytes pointer. Issue was reported and fixed by @TheLX5.