Remove the ability to disable the 255 sprites per level patch.
This commit fixes an oversight with Pixi 1.40, where the minor sprites tables (smoke, minor extended, bounce, etc) use RAM addresses in the $1938-$19B7 range unconditionally without checking if the 255 sprites per level patch was disabled or not. The $1938 table is only free when the patch is in effect (which is by default with pixi), so this means that when the -d255spl flag is passed when pixi is ran those tables point to RAM which is already in use by the game. Another possible fix is to make the the tables use other freeram instead of $1938 but that would require reserving even more RAM in the $7Fxxxx range which would be unwise imo. I also took this path to fixing the issue because if you apply pixi normally it applies some hijacks to properly work with the 255 sprites per level patch which don't get remove if you then reapply pixi with -d255spl, which means that the flag is already half-implemented as-is, thus removing support completely is better.
For posterity I'm not removing any code related to the flag, I just made it so Pixi would error immediately when it's passed.